backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
langchain
langchain-openai
langchain-community
chromadb
faiss-cpu
pymupdf
python-docx
apscheduler
python-dotenv
pandas
pydantic-settings
